# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pixelate",
        "description": "Documentation for Pixelate app",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Server"
        }
    ],
    "paths": {
...
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get Post by Id",
                "description": "Get Single Post",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Post"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Display the specified Post.\nGET|HEAD /posts/{id}",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Display the specified Post.\nGET|HEAD /posts/{id}",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update Post",
                "description": "Update A Post by Id",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass Post data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                        "type": "string",
                                        "example": "Here"
                                    },
                                    "location": {
                                        "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                        "type": "string",
                                        "example": "Dusseldorf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "Post deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Update the specified Post in storage.\nPUT/PATCH /posts/{id}",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete Post",
                "description": "Delete A Post by Id",
                "operationId": "App\\Http\\Controllers\\API\\PostAPIController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "message": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "Post deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "description": "Remove the specified Post from storage.\nDELETE /posts/{id}",
                                            "type": "string",
                                            "example": "Post not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/designs": {
            "get": {
                "tags": [
                    "Design"
                ],
                "summary": "Get designs that are Live",
                "description": "Get all designs available online (set to live )",
                "operationId": "App\\Http\\Controllers\\Designs\\DesignController::index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Design"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UDesign Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Design"
                ],
                "summary": "Upload Image",
                "description": "Uploads Image Local or to S3 bucket",
                "operationId": "App\\Http\\Controllers\\Designs\\UploadController::upload",
                "requestBody": {
                    "description": "Upload images request body",
                    "required": true,
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "type": "string",
                                "format": "binary"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Design"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "image": {
                                                    "example": [
                                                        "The image field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/designs/all": {
            "get": {
                "tags": [
                    "Design"
                ],
                "summary": "Get all designs",
                "description": "Get all designs available",
                "operationId": "App\\Http\\Controllers\\Designs\\DesignController::allDesigns",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Design"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "UDesign Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comment": {
                "properties": {
                    "id": {
                        "description": "App\\Models\\Comment",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_dates": {
                        "description": "App\\Models\\Comment",
                        "properties": {
                            "created_at_human": {
                                "description": "Date Created formatted",
                                "example": "52 minutes ago"
                            },
                            "created_at": {
                                "description": "Raw unfarmatted Date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "updated_dates": {
                        "description": "App\\Models\\Comment",
                        "properties": {
                            "updated_at_human": {
                                "description": "Date Updated formatted",
                                "example": "52 minutes ago"
                            },
                            "updated_at": {
                                "description": "Raw unfarmatted update date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Comment"
                }
            },
            "Design": {
                "properties": {
                    "id": {
                        "description": "App\\Models\\Design",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "description": "App\\Models\\Design",
                        "type": "string",
                        "example": "the sea life"
                    },
                    "slug": {
                        "description": "App\\Models\\Design",
                        "type": "string",
                        "example": "the-sea-life"
                    },
                    "disk": {
                        "description": "App\\Models\\Design",
                        "type": "string",
                        "example": "public"
                    },
                    "is_live": {
                        "description": "App\\Models\\Design",
                        "type": "boolean",
                        "example": 0
                    },
                    "images": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "thumbnail": {
                                "description": "App\\Models\\Design",
                                "type": "string",
                                "example": "http://localhost:8000/storage/uploads/designs/thumbnail/"
                            },
                            "large": {
                                "description": "App\\Models\\Design",
                                "type": "string",
                                "example": "http://localhost:8000/storage/uploads/designs/large/"
                            },
                            "original": {
                                "description": "App\\Models\\Design",
                                "type": "string",
                                "example": "http://localhost:8000/storage/uploads/designs/original/"
                            }
                        },
                        "type": "object"
                    },
                    "likes_count": {
                        "description": "App\\Models\\Design",
                        "type": "integer",
                        "example": 0
                    },
                    "uploaded_successful": {
                        "description": "App\\Models\\Design",
                        "type": "interger",
                        "example": null
                    },
                    "tag_list": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "tag": {
                                "description": "List of Tags added to design, array of many tags",
                                "example": [
                                    "Cool",
                                    "Bright Day"
                                ]
                            },
                            "normalized": {
                                "description": "List of Tags normalized, lowercase and slugged",
                                "example": [
                                    "cool",
                                    "bright-day"
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "created_dates": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "created_at_human": {
                                "description": "Date Created formatted",
                                "example": "52 minutes ago"
                            },
                            "created_at": {
                                "description": "Raw unfarmatted Date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "updated_dates": {
                        "description": "App\\Models\\Design",
                        "properties": {
                            "updated_at_human": {
                                "description": "Date Updated formatted",
                                "example": "52 minutes ago"
                            },
                            "updated_at": {
                                "description": "Raw unfarmatted update date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "comment": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Design"
                }
            },
            "Post": {
                "properties": {
                    "id": {
                        "description": "Class Post",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "caption": {
                        "description": "Class Post",
                        "type": "string",
                        "example": "the sea life"
                    },
                    "location": {
                        "description": "Class Post",
                        "type": "string",
                        "example": "Dusseldorf"
                    },
                    "imageUrl": {
                        "description": "Class Post",
                        "type": "string",
                        "example": "http://localhost:8000/storage/3B5wcGnEbnFaAYsjMxk5P1V1fAGKaVhviC3EO0Gd.png"
                    },
                    "created_dates": {
                        "description": "Class Post",
                        "properties": {
                            "created_at_human": {
                                "description": "Date Created formatted",
                                "example": "52 minutes ago"
                            },
                            "created_at": {
                                "description": "Raw unfarmatted Date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "updated_dates": {
                        "description": "Class Post",
                        "properties": {
                            "updated_at_human": {
                                "description": "Date Updated formatted",
                                "example": "52 minutes ago"
                            },
                            "updated_at": {
                                "description": "Raw unfarmatted update date ",
                                "example": "2020-11-09T20:04:11.000000Z"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "comment": {
                        "$ref": "#/components/schemas/Comment"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
...
        },
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Example of Auth Header Token (Copy and paste below to authorize)=>\nBearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9sb2dpbiIsImlhdCI6MTYwNTM0ODU1OSwiZXhwIjoxNjA1MzUyMTU5LCJuYmYiOjE2MDUzNDg1NTksImp0aSI6IjlqMDN0MVVISThoNjlhRUsiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.dn45EdZMatpW2zAQcZklDS2902mwVNx44TbkDiiGLXY",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}